Remove incorrect blob retrieval
NM_SETTING_802_1X_CK_SCHEME_BLOB scheme requires the file path to set the blob, but returns the blob itself on retrieval.
API was incorrectly treating this blob as a string (converting to UTF8 as best effort) and storing it in the public facing cert file path buffer.

Remove the retrieval logic and rename gbytes_to_string() to ssid_gbytes_to_string() to more clearly indicate this API can only be used with SSIDs.

Bug: 19661