lrd-userspace-examples: swclient: Add functions to open/read/close the progress IPC channel
This change adds proper Python access to opening, reading, and closing the
progress messages via the IPC channel. The intention is for a Python application
to call the functions in the proper order (i.e., open, read, close).

The read() function can be wrapped by the Python select.poll() function to allow
monitoring of a firmware update progress.

Also included are various code cleanups and bug fixes including the removal of
the get_fw_update_state() function (this functionality is now replaced with
the open/read/close functions).

Bug 21723